git-svn-id: file:///semplestb/final_repo@1121 02200ff9-b5b2-46f0-9171-221b09c08c7b

+++ SEMPlestDBScripts/StoredProcs/AddKeywordBidData.sql
@@ -12,7 +12,7 @@ GO
 CREATE PROCEDURE dbo.AddKeywordBidData
 (
 
-	@PromotionID        BIGINT,
+	@PromotionID			INT,
 	@Keyword				NVARCHAR(250),
 	@AdvertisingEngine		VARCHAR(50),
 	@BidType				VARCHAR(25),
